+++ shared/types/APIService.ts
